fix(layout stacking bug): remove overriding fetch from stacking the panel group
This commit fixes the bug that made the Resizable Panel Group throw errors, it was caused by a !flex on the panel group, this was found by taking a look into an, until the time of this commit, unmerged pr into to the stacking-order package

The pr aims to fix the problem related to flex with z-index. Until it's merged, the solution is to not override the css directives related to that in this code.